Wired up local type inference.
Also split out params and constraints in the AST. They're still specified via
the same syntactic mechanism, but keeping them in the same list was causing
problems.

Also also moved array index operator parsing into the binop hierarchy. I was
originally parsing it like record select but that's not valid because the right
hand side is a complete expression, not just an identifier. So it needs to be
parsed like any other binop. I should probably just turn it into a funapply
like the rest of the binops.