Constructor funs, recursive unfolding, syms in types.
- We now have symbols for record constructor functions (and singleton values),
  so an expression like `Cons(1, Nil)` now gets a correct type.

- Types now refer to their defining symbol rather than its name. This allows us
  to correctly (modulo other errors of incomprehension) join union cases to the
  union type, for example. And in general it's clearly better. A type
  definition yields a unique symbol, that symbol is a better placeholder for
  the type than a name.

- Recursive types are handled a bit better. When we give a tree a record field
  type, we unfold that type once to get out from under a potential lazy
  (recursive) reference.

DefTree is also now supplied (during indexing) with a reference to the symbol
created for the definition it provides. I may also introduce a RefTree at some
point which will host a reference to the referent symbol, though in theory one
would have the context needed to just look it up in the symbol table... TBD.

This also changes (again) the way symbols yield their types. The type of a
symbol once again comes from an abstract method instead of an "infoFn" supplied
to the Symbol constructor. Because types and symbols are now mutually referent,
this is needed to cut the Gordian knot when defining primitive types. There we
want to just define the symbol and the type all at once, rather than by way of
an AST fragment. There's no AST for declaring primitives, something I may end
up changing.