Various progress on impl resolution.
Also a bunch of revamping along the way. We now resolve simple interface
methods to their appropriate implementations during typing.

DefTree now has a type signature which is separate from the typed tree. This
types just the signature parts of the tree but not the bodies. This allows us
to type the signatures of all defs before typing any bodies, avoiding further
issues with recursion. Symbols (which represent def trees) use the signature
now rather than the tree type.