Started to work on lowering.
I started down the road of generating JS directly, but the mismatch in language
constructs made it very cumbersome. So even if I only ever generate JS from C0,
it will be easier to do it from a lowered AST.

I'm kind of wishing I was generating JVM bytecode instead of JS because that'd
be a easier in many cases from these lowered trees. With JS I have to rename
variables in nested blocks and handle lowering of blocks in inconvenient places
(like the condition of a while loop). I suppose with bytecode I'd have to deal
with extracting nested functions and other PITAs, so neither target is the
"easy one".