Revamp lowering somewhat.
Previously I was doing the simple thing: pushing everything out practically
into SSA form such that everything was bound to a name and then fun args were
just names. That was convenient for turning "everything is an expression" into
"my sad target language only supports expressions in limited contexts" but it
made for some really verbose and ugly generated JavaScript.

Now we thread a notion of "binding target" through the lowering process, so
when we need to hoist something into a binding we can (like when lowering an if
expression to a binding plus if statement, ditto for block expressions, match
expressions, etc.), but when we don't have to hoist, we just let things remain
inline.

There's an annoying niggle: if any funarg needs hoisting we have to hoist all
the earlier funargs too, which complicates the process of deciding whether
hoisting is needed in general. For now I've punted and have a separate tree
traversal that checks whether we need to hoist. In the case of funargs, if any
funarg needs hoisting they all get hoisted (regardless of whether some number
of args on the end of the list could remain unhoisted).