Various progress on editing.
The markup tree is annotated with editing functions which indicate what changes
to make to the AST when a markup element is edited. The updated AST is then
reformatted and the new markup displayed.

Also, I have unfortunately reached my bullshit limit with the PureScript
ecosystem (especially, though not limited to, Halogen). So I'm going to take
another crack at Reason, and if that proves to be equally bullshit laden, I'm
just going to write the goddamned first editor and compiler in Flow annotated
JavaScript. Sigh.