Have basic editing mostly working.
I think this catches us up to where we were in PureScript land. Now we can move
forward, yay!

I don't even have to deal with the straw that broke this camel's back in
PureScript (dynamically assigning focus to a text input) because React seems to
destroy and recreate the text input between edits, so the autofocus attribute
does what is needed.

Also, I'm already liking MOBX. It's very much in the same spirit that I had in
mind for RSP: fine grained external reactive state, views are generated from
projections of that state. Using it with React is not quite two way data
binding, but that would probably be more trouble than it's worth anyway (since
React is designed around one way data binding).