Major tree revamp; progress on symbols, types.
Trees are no longer represented with subclasses of Tree. Now we have a generic
Tree node with arbitrary branches and we just have to "know" the shape of the
node we're looking at/manipulating.

This sucks from a static typing perspective and it's easy to screw up, but I
couldn't handle all the code duplication in the dozen plus tree nodes and it
was only getting worse (more node types, more duplicated code per node). Maybe
I'll come up with a way to preserve some modicum of safety but for now we're
going to dance with this devil instead of the other one.

This also introduces symbols and a notion of a symbol table, and tree types
(different from type trees), though we're not anywhere near doing type checking
or inference yet.

It's still somewhat in transition. I mainly only get to work on this M-W, so I
want to commit (and comment on) it before ignoring it for four days.