Syntax, span changes, completions, etc.
Hastily undid terrifying C-like syntax experiment with all them parens and
square and curly brackets. We're back to a Haskelly syntax with inline type
ascriptions.

Span is now a full fledged class which embeds the editor functionality. This
will allow us to more carefully coordinate between the text shown and the
actual logical entity underlying it, which is especially needed for
completions.

Added very simple symbol completion for type/term ref editing. Much work still
to be done on this front.