Revamped pattern trees, more annotation work.
I was trying to structure pattern trees like let bindings such that the body of
the pattern was "under" the binding tree, but that's not going to work for
non-trivial patterns and introduces all sorts of complications to typing
pattern trees.

So we have a 'case' tree again with a pattern subtree and a body subtree, and
the pattern subtree can be typed in a fairly sensible way. The case introduces
a special (currently very inefficient) scope over the body which binds all
symbols introduced by the pattern tree.

Typing unapplies is presently kind of crazy. I'm not sure if this approach will
stand the test of time, but YOLO!

Also added UI toggle for showing types, and did some work on widening constants
to scalars.