Completion & edit action work.
We now detect literals during the completion phase and return a completion that
inserts the literal into the tree. Completions now do their own tree edits as
well instead of relying on the span editing code to handle that.

Span editing has been torn apart and turned into a more rule based system,
where we predicate particular edit actions on a combination of a key press and
a path expression. This will undoubtedly need to get more sophisticated, but it
allows the myriad (and growing) different special snowflake actions that take
place at different points in the tree to be largely separated from one another.