A bunch of work on editing &c.
Started adding support for inserting holes in a direction based on the current
cursor. You can add a hole to the left which makes the current node an arg in a
funapp, or you can add a hole above or below a case branch, adding another case
to a match. This will surely evolve, but it's a start.

Also more fiddling with formatting, what values are shown by default,
automatically starting to edit holes when you navigate to them, etc.