Allow selecting spans with the mouse.
In addition to wiring up onClick handlers, we needed to invert the way we were
tracking the selected span.

We used to peel off the head of the cursor as we rendered spans and learned
that a span was selected if it was at the "end" of the cursor. Now we build up
the path to each span we're rendering and compare it with the full path
specified by the cursor to determine if a span is selected.

This allows us to use the built-up path at each span in the onClick handler to
simply say "set the cursor to this path". Less weird, more useful.