Starting down road to undo/redo & tree serialize.
Was just doing undo/redo but then realized that my hacky test tree creation
code was going to get really annoying to maintain. Decided to convert that to
unserializing hand serialized trees so that I would not be erroneously inclined
to preserve the ability to create trees manually during the undo/redo revamps.

Everything is still in progress, but we're back to the point of being able to
view and interactively create trees, as well/poorly as before anyway.