Reinstate support for creating new defs.
This uses a different UI modality because the 'def hole' was turning into too
much of a hack. We really need to know what kind of def we're creating before
we create any sort of editable tree over in the stack.

I will probably change this yet again if/when we move to a 2D grid of
characters "UI" (even though we'll still be editing trees), but for now this
lets us do what we need without weirding the code too much.