Spans can now respond to keys when not editing.
Wired up action where pressing space on a field def inserts a new field def at
the cursor.

This may go away or change if we switch to a modeless editing model.

Also added Tree.relink() for when a branch is just moving rather than being
unlinked and relinked. This avoids extra work and id churn when we want to
insert/remove a branch in a variable arity tree.