Various work on key bindings &c.
We now convert all presses to 'key chords' and look up rules based on explicit
comparison with chords and path matching. This will allow us shortly to display
a list of all active key bindings (and a heuristic description of what they'll
do), updated on the fly based on the current cursor position.