More progress on unified editing.
Things are still messy, but starting to congeal around something sane. All key
handling is now done by the reactive data model, the rendering part (the React
components) just display whatever is in the store. Simpler, cleaner, better.

We no longer have a SpanStore for the currently edited span. We're always
potentially editing. Info on the currently selected span is tracked in
DefStore, and we "prepare" to edit every span as it is selected. Then if any
action is taken that initiates an edit, we transition to text edit mode (still
rough edges in this area). If any action is taken that indicates a tree
navigation, we exit text edit mode.