A bunch of signature computation fiddling.
This is semi-incomplete, but I'm committing it because I think I'm about to
tear it all out and redo typing to be more "traditional" in that we'll do
typing in a single pass over a definition and generate a mapping from tree node
to type (using an algorithm like complete and easy bidir), and then recompute
any time the tree changes.

This twisty interconnected dataflow approach is proving to be too fiddly, and
is probably a bad idea to start with, at least without first sinking a bunch of
time into research and theory, instead of just ad-hoc-ing it together.