More conversion to bidir typer.
A lot of things are working now, but I have to go back to the drawing board on
how to edit type signatures. I still like having them intertwingled with their
associated terms, i.e.:

  reverse ∀A as:List A → List A = ...

instead of:

  reverse :: ∀A List A → List A
  reverse as = ...

but I'm not sure whether the same editing experience I had before is ideal.
More thinking to be done...