Various progress on parsing, typing.
We can now parse and type check things like this:

  def fst :: ∀A ⇒ ∀B ⇒ x:A → y:B → A = x

Which also represents my latest thinking on (interim) signature syntax. I've
gone back and forth about a zillion times, but I like a few things about this.

The arrows are syntactically part of the argument that precedes them, so a zero
argument definition omits the arguments and arrows and ends up with just a type
ascription:

  def zero :: Int = 0

The double colon is only used for top-level ascriptions. "Named fun args" are
only part of a top-level definition ascription as well. So a higher-order
function signature would look like:

  def foldl :: ∀F ⇒ Foldable F ⇒ ∀A ⇒ ∀Z ⇒
               z:Z → as:F A → f:(Z → A → Z) → Z = ...

That also shows type class constraints, which are intermixed with type
arguments & also use the fat arrow. (Ignore the fact that foldl would probably
be a member of the Foldable type class.) Fat arrow indicates implict arguments
and skinny arrows explicit, rather than Haskell's interpretation of fat arrow
as logical implication. I'm taking a more operational view.

Anyhow, I'm not sure how much of this will change when syntax no longer needs
to be parseable by a computer, but I like this temporary syntax better than
what I've managed in the past.